TMP.TBOT/requirements.txt
beautifulsoup4==4.12.2
jdatetime==4.1.1
python-telegram-bot==20.5
Requests==2.31.0
schedule==1.2.0
urllib3==1.25.8
